ironclad/requirements.txt
Flask
torch
torchvision
facenet-pytorch
PIL
faiss-cpu
scikit-learn
matplotlib
seaborn
numpy